requirement.txt
SQLAlchemy==1.3.19
Jinja2==2.11.3
unicorn==1.0.2
uvicorn==0.13.3
python-dotenv==0.15.0
Jinja2==2.11.3
fastapi==0.63.0
aiofiles==0.6.0
mysql-connector==2.2.9
pytest==5.3.5
requests==2.24.0